makes business_days and business_hours comparable.
But not to each other, that is a larger and probably impossible problem.
For instance, if you wanted to see if 21.business_hours is less than
3.business_days, you'd need to know how many hours are in a business day.
And since each day can have a different number of hours (imagine half-day
Saturdays), the question probably isn't pre-answerable without providing
a date to calculate it in reference to.